Server patch in description
Fixed a bug where it didn't allow to send a big file (tried 11mb). now it works using non blocking send and select.

Also, added MANY "state" prints that the server does to show what it's doing right now. imo it's useful.

also added more comments and fixed a little bug where it crashed the server if a user tried to register with a name that's already available.